afxdp/tunnel: relocate inline tests to tunnel_tests.rs
Same playbook. tunnel.rs has prod code on both sides of the inline test
block (lines 1-410, then 411-534 at base) — splice handled by relocate_v2.
After: tunnel.rs drops from 535 → 432 LOC.
cargo build/test 92/878.
